fix(Time): 예약이 존재하는 Time 삭제시 예외 처리 수정

+++ src/main/java/roomescape/service/TimeService.java
@@ -45,6 +45,10 @@ public class TimeService {
 
     public void removeTime(long timeId) {
         getTimeOrElseThrow(timeId);
+        if (timeRepository.existsByReferencedId(timeId)) {
+            throw new CustomException(ErrorCode.TIME_IS_REFERENCED);
+        }
+
         timeRepository.deleteById(timeId);
     }
